Extract the atiny_adapter user API, implemented in sys_init.c
