+++ targets/STM32F429IGTx_FIRE/Src/sys_init.c
@@ -172,7 +172,21 @@ void hieth_hw_init(void)
     (void)LOS_HwiCreate(ETH_IRQn, 1,0,ETH_IRQHandler,0);
 }
 
-void sys_reboot(void)
+/**
+ * atiny_adapter user interface 
+ */
+void atiny_usleep(unsigned long usec)
+{
+    delayus((uint32_t)usec);
+}
+
+int atiny_random(void* output, size_t len)
+{
+    return hal_rng_generate_buffer(output, len);
+}
+
+void atiny_reboot(void)
 {
     HAL_NVIC_SystemReset();
 }
+
